Port user tests from test_models
