+++ tests/ignore/test_models.py
@@ -1,32 +1,22 @@
 # -*- coding: utf-8 -*-
 '''Unit tests for models and their factories.'''
 import datetime
-import json
-import os.path
 import random
 import string
 import unittest
 import urllib
-import urlparse
 
-import itsdangerous
 import mock
 import pytz
 from dateutil import parser
 from django.utils import timezone
-from framework.analytics import get_total_activity_count
-from framework.auth import exceptions as auth_exc
 from framework.auth import Auth, User, cas
-from framework.auth.exceptions import ChangePasswordError, ExpiredTokenError
-from framework.auth.signals import user_merged
-from framework.auth.utils import impute_names_model
-from framework.bcrypt import check_password_hash
 from framework.celery_tasks import handlers
 from framework.exceptions import PermissionsError
 from framework.sessions import set_session
-from framework.sessions.model import Session
 from modularodm import Q
-from modularodm.exceptions import ValidationError, ValidationValueError
+#from modularodm.exceptions import ValidationError, ValidationValueError
+from osf.exceptions import ValidationError, ValidationValueError
 from nose.tools import *  # noqa (PEP8 asserts)
 from tests.base import (Guid, OsfTestCase, capture_signals, fake,
                         get_default_metaschema)
@@ -41,7 +31,7 @@ from tests.factories import (ApiOAuth2ApplicationFactory, AuthUserFactory,
                              SessionFactory, UnconfirmedUserFactory,
                              UnregUserFactory, UserFactory, WatchConfigFactory)
 from tests.utils import mock_archive
-from website import filters, language, mailchimp_utils, settings
+from website import language, settings
 from addons.wiki.exceptions import (NameEmptyError, NameInvalidError,
                                             NameMaximumLengthError,
                                             PageCannotRenameError,
@@ -63,991 +53,6 @@ from website.project.sanctions import Sanction, DraftRegistrationApproval
 from website.views import find_bookmark_collection
 
 GUID_FACTORIES = UserFactory, NodeFactory, ProjectFactory
-
-
-class TestUserValidation(OsfTestCase):
-
-    def setUp(self):
-        super(TestUserValidation, self).setUp()
-        self.user = AuthUserFactory()
-
-    def test_validate_fullname_none(self):
-        self.user.fullname = None
-        with assert_raises(ValidationError):
-            self.user.save()
-
-    def test_validate_fullname_empty(self):
-        self.user.fullname = ''
-        with assert_raises(ValidationValueError):
-            self.user.save()
-
-    def test_validate_social_profile_websites_empty(self):
-        self.user.social = {'profileWebsites': []}
-        self.user.save()
-        assert_equal(self.user.social['profileWebsites'], [])
-
-    def test_validate_social_profile_website_many_different(self):
-        basepath = os.path.dirname(__file__)
-        url_data_path = os.path.join(basepath, '../website/static/urlValidatorTest.json')
-        with open(url_data_path) as url_test_data:
-            data = json.load(url_test_data)
-
-        fails_at_end = False
-        for should_pass in data["testsPositive"]:
-            try:
-                self.user.social = {'profileWebsites': [should_pass]}
-                self.user.save()
-                assert_equal(self.user.social['profileWebsites'], [should_pass])
-            except ValidationError:
-                fails_at_end = True
-                print('\"' + should_pass + '\" failed but should have passed while testing that the validator ' + data['testsPositive'][should_pass])
-
-        for should_fail in data["testsNegative"]:
-            self.user.social = {'profileWebsites': [should_fail]}
-            try:
-                with assert_raises(ValidationError):
-                    self.user.save()
-            except AssertionError:
-                fails_at_end = True
-                print('\"' + should_fail + '\" passed but should have failed while testing that the validator ' + data['testsNegative'][should_fail])
-        if fails_at_end:
-            raise
-
-    def test_validate_multiple_profile_websites_valid(self):
-        self.user.social = {'profileWebsites': ['http://cos.io/', 'http://thebuckstopshere.com', 'http://dinosaurs.com']}
-        self.user.save()
-        assert_equal(self.user.social['profileWebsites'], ['http://cos.io/', 'http://thebuckstopshere.com', 'http://dinosaurs.com'])
-
-    def test_validate_social_profile_websites_invalid(self):
-        self.user.social = {'profileWebsites': ['help computer']}
-        with assert_raises(ValidationError):
-            self.user.save()
-
-    def test_validate_multiple_profile_social_profile_websites_invalid(self):
-        self.user.social = {'profileWebsites': ['http://cos.io/', 'help computer', 'http://dinosaurs.com']}
-        with assert_raises(ValidationError):
-            self.user.save()
-
-    def test_empty_social_links(self):
-        assert_equal(self.user.social_links, {})
-        assert_equal(len(self.user.social_links), 0)
-
-    def test_profile_website_unchanged(self):
-        self.user.social = {'profileWebsites': ['http://cos.io/']}
-        self.user.save()
-        assert_equal(self.user.social_links['profileWebsites'], ['http://cos.io/'])
-        assert_equal(len(self.user.social_links), 1)
-
-    def test_various_social_handles(self):
-        self.user.social = {
-            'profileWebsites': ['http://cos.io/'],
-            'twitter': 'OSFramework',
-            'github': 'CenterForOpenScience'
-        }
-        self.user.save()
-        assert_equal(self.user.social_links, {
-            'profileWebsites': ['http://cos.io/'],
-            'twitter': 'http://twitter.com/OSFramework',
-            'github': 'http://github.com/CenterForOpenScience'
-        })
-
-    def test_multiple_profile_websites(self):
-        self.user.social = {
-            'profileWebsites': ['http://cos.io/', 'http://thebuckstopshere.com', 'http://dinosaurs.com'],
-            'twitter': 'OSFramework',
-            'github': 'CenterForOpenScience'
-        }
-        self.user.save()
-        assert_equal(self.user.social_links, {
-            'profileWebsites': ['http://cos.io/', 'http://thebuckstopshere.com', 'http://dinosaurs.com'],
-            'twitter': 'http://twitter.com/OSFramework',
-            'github': 'http://github.com/CenterForOpenScience'
-        })
-
-    def test_nonsocial_ignored(self):
-        self.user.social = {
-            'foo': 'bar',
-        }
-        self.user.save()
-        assert_equal(self.user.social_links, {})
-
-    def test_validate_jobs_valid(self):
-        self.user.jobs = [{
-            'institution': 'School of Lover Boys',
-            'department': 'Fancy Patter',
-            'title': 'Lover Boy',
-            'startMonth': 1,
-            'startYear': '1970',
-            'endMonth': 1,
-            'endYear': '1980',
-        }]
-        self.user.save()
-
-    def test_validate_jobs_institution_empty(self):
-        self.user.jobs = [{'institution': ''}]
-        with assert_raises(ValidationError):
-            self.user.save()
-
-    def test_validate_jobs_bad_end_date(self):
-        # end year is < start year
-        self.user.jobs = [{
-            'institution': fake.company(),
-            'department': fake.bs(),
-            'position': fake.catch_phrase(),
-            'startMonth': 1,
-            'startYear': '1970',
-            'endMonth': 1,
-            'endYear': '1960',
-        }]
-        with assert_raises(ValidationValueError):
-            self.user.save()
-
-    def test_validate_schools_bad_end_date(self):
-        # end year is < start year
-        self.user.schools = [{
-            'degree': fake.catch_phrase(),
-            'institution': fake.company(),
-            'department': fake.bs(),
-            'startMonth': 1,
-            'startYear': '1970',
-            'endMonth': 1,
-            'endYear': '1960',
-        }]
-        with assert_raises(ValidationValueError):
-            self.user.save()
-
-    def test_validate_jobs_bad_year(self):
-        start_year = ['hi', '20507', '99', '67.34']
-        for year in start_year:
-            self.user.jobs = [{
-                'institution': fake.company(),
-                'department': fake.bs(),
-                'position': fake.catch_phrase(),
-                'startMonth': 1,
-                'startYear': year,
-                'endMonth': 1,
-                'endYear': '1960',
-            }]
-            with assert_raises(ValidationValueError):
-                self.user.save()
-
-    def test_validate_schools_bad_year(self):
-        start_year = ['hi', '20507', '99', '67.34']
-        for year in start_year:
-            self.user.schools = [{
-                'degree': fake.catch_phrase(),
-                'institution': fake.company(),
-                'department': fake.bs(),
-                'startMonth': 1,
-                'startYear': year,
-                'endMonth': 1,
-                'endYear': '1960',
-            }]
-            with assert_raises(ValidationValueError):
-                self.user.save()
-
-
-class TestUser(OsfTestCase):
-
-    def setUp(self):
-        super(TestUser, self).setUp()
-        self.user = UserFactory()
-        self.auth = Auth(user=self.user)
-
-    def test_repr(self):
-        assert_in(self.user.username, repr(self.user))
-        assert_in(self.user._id, repr(self.user))
-
-    def test_update_guessed_names(self):
-        name = fake.name()
-        u = User(fullname=name)
-        u.update_guessed_names()
-        u.save()
-
-        parsed = impute_names_model(name)
-        assert_equal(u.fullname, name)
-        assert_equal(u.given_name, parsed['given_name'])
-        assert_equal(u.middle_names, parsed['middle_names'])
-        assert_equal(u.family_name, parsed['family_name'])
-        assert_equal(u.suffix, parsed['suffix'])
-
-    def test_non_registered_user_is_not_active(self):
-        u = User(username=fake.email(),
-                 fullname='Freddie Mercury',
-                 is_registered=False)
-        u.set_password('killerqueen')
-        u.save()
-        assert_false(u.is_active)
-
-    def test_create_unregistered(self):
-        name, email = fake.name(), fake.email()
-        u = User.create_unregistered(email=email,
-                                     fullname=name)
-        u.save()
-        assert_equal(u.username, email)
-        assert_false(u.is_registered)
-        assert_false(u.is_claimed)
-        assert_true(u.is_invited)
-        assert_false(email in u.emails)
-        parsed = impute_names_model(name)
-        assert_equal(u.given_name, parsed['given_name'])
-
-    @mock.patch('framework.auth.core.User.update_search')
-    def test_search_not_updated_for_unreg_users(self, update_search):
-        u = User.create_unregistered(fullname=fake.name(), email=fake.email())
-        u.save()
-        assert not update_search.called
-
-    @mock.patch('framework.auth.core.User.update_search')
-    def test_search_updated_for_registered_users(self, update_search):
-        UserFactory(is_registered=True)
-        assert_true(update_search.called)
-
-    def test_create_unregistered_raises_error_if_already_in_db(self):
-        u = UnregUserFactory()
-        dupe = User.create_unregistered(fullname=fake.name(), email=u.username)
-        with assert_raises(ValidationValueError):
-            dupe.save()
-
-    def test_user_with_no_password_is_not_active(self):
-        u = User(
-            username=fake.email(),
-            fullname='Freddie Mercury',
-            is_registered=True,
-        )
-        u.save()
-        assert_false(u.is_active)
-
-    def test_merged_user_is_not_active(self):
-        master = UserFactory()
-        dupe = UserFactory(merged_by=master)
-        assert_false(dupe.is_active)
-
-    def test_merged_user_with_two_account_on_same_project_with_different_visibility_and_permissions(self):
-        user2 = UserFactory.build()
-        user2.save()
-
-        project = ProjectFactory(is_public=True)
-        # Both the master and dupe are contributors
-        project.add_contributor(user2, log=False)
-        project.add_contributor(self.user, log=False)
-        project.set_permissions(user=self.user, permissions=['read'])
-        project.set_permissions(user=user2, permissions=['read', 'write', 'admin'])
-        project.set_visible(user=self.user, visible=False)
-        project.set_visible(user=user2, visible=True)
-        project.save()
-        self.user.merge_user(user2)
-        self.user.save()
-        project.reload()
-        assert_true('admin' in project.permissions[self.user._id])
-        assert_true(self.user._id in project.visible_contributor_ids)
-        assert_false(project.is_contributor(user2))
-
-    def test_cant_create_user_without_username(self):
-        u = User()  # No username given
-        with assert_raises(ValidationError):
-            u.save()
-
-    def test_date_registered_upon_saving(self):
-        u = User(username=fake.email(), fullname='Foo bar')
-        u.save()
-        assert_true(u.date_registered)
-
-    def test_create(self):
-        name, email = fake.name(), fake.email()
-        user = User.create(
-            username=email, password='foobar', fullname=name
-        )
-        user.save()
-        assert_true(user.check_password('foobar'))
-        assert_true(user._id)
-        assert_equal(user.given_name, impute_names_model(name)['given_name'])
-
-    def test_create_unconfirmed(self):
-        name, email = fake.name(), fake.email()
-        user = User.create_unconfirmed(
-            username=email, password='foobar', fullname=name
-        )
-        user.save()
-        assert_false(user.is_registered)
-        assert_equal(len(user.email_verifications.keys()), 1)
-        assert_equal(
-            len(user.emails),
-            0,
-            'primary email has not been added to emails list'
-        )
-
-    def test_create_unconfirmed_from_external_service(self):
-        name, email = fake.name(), fake.email()
-        external_identity = {
-            'ORCID': {
-                fake.ean(): 'CREATE'
-            }
-        }
-        user = User.create_unconfirmed(
-            username=email,
-            password=str(fake.password()),
-            fullname=name,
-            external_identity=external_identity,
-        )
-        user.save()
-        assert_false(user.is_registered)
-        assert_equal(len(user.email_verifications.keys()), 1)
-        assert_equal(user.email_verifications.popitem()[1]['external_identity'], external_identity)
-        assert_equal(
-            len(user.emails),
-            0,
-            'primary email has not been added to emails list'
-        )
-
-    def test_create_confirmed(self):
-        name, email = fake.name(), fake.email()
-        user = User.create_confirmed(
-            username=email, password='foobar', fullname=name
-        )
-        user.save()
-        assert_true(user.is_registered)
-        assert_true(user.is_claimed)
-        assert_equal(user.date_registered, user.date_confirmed)
-
-    def test_cant_create_user_without_full_name(self):
-        u = User(username=fake.email())
-        with assert_raises(ValidationError):
-            u.save()
-
-    @mock.patch('website.security.random_string')
-    def test_add_unconfirmed_email(self, random_string):
-        token = fake.lexify('???????')
-        random_string.return_value = token
-        u = UserFactory()
-        assert_equal(len(u.email_verifications.keys()), 0)
-        u.add_unconfirmed_email('foo@bar.com')
-        assert_equal(len(u.email_verifications.keys()), 1)
-        assert_equal(u.email_verifications[token]['email'], 'foo@bar.com')
-
-    @mock.patch('website.security.random_string')
-    def test_add_unconfirmed_email_adds_expiration_date(self, random_string):
-        token = fake.lexify('???????')
-        random_string.return_value = token
-        u = UserFactory()
-        u.add_unconfirmed_email("test@osf.io")
-        assert_is_instance(u.email_verifications[token]['expiration'], datetime.datetime)
-
-    def test_add_blank_unconfirmed_email(self):
-        with assert_raises(ValidationError) as exc_info:
-            self.user.add_unconfirmed_email('')
-        assert_equal(exc_info.exception.message, "Invalid Email")
-
-    def test_add_blacklisted_domain_unconfirmed_email(self):
-        with assert_raises(ValidationError) as e:
-            self.user.add_unconfirmed_email('kanye@mailinator.com')
-        assert_equal(e.exception.message, 'Invalid Email')
-
-    @mock.patch('website.security.random_string')
-    def test_get_confirmation_token(self, random_string):
-        random_string.return_value = '12345'
-        u = UserFactory()
-        u.add_unconfirmed_email('foo@bar.com')
-        assert_equal(u.get_confirmation_token('foo@bar.com'), '12345')
-        assert_equal(u.get_confirmation_token('fOo@bar.com'), '12345')
-
-    def test_get_confirmation_token_when_token_is_expired_raises_error(self):
-        u = UserFactory()
-        # Make sure token is already expired
-        expiration = timezone.now() - datetime.timedelta(seconds=1)
-        u.add_unconfirmed_email('foo@bar.com', expiration=expiration)
-
-        with assert_raises(ExpiredTokenError):
-            u.get_confirmation_token('foo@bar.com')
-
-    @mock.patch('website.security.random_string')
-    def test_get_confirmation_token_when_token_is_expired_force(self, random_string):
-        random_string.return_value = '12345'
-        u = UserFactory()
-        # Make sure token is already expired
-        expiration = timezone.now() - datetime.timedelta(seconds=1)
-        u.add_unconfirmed_email('foo@bar.com', expiration=expiration)
-
-        # sanity check
-        with assert_raises(ExpiredTokenError):
-            u.get_confirmation_token('foo@bar.com')
-
-        random_string.return_value = '54321'
-
-        token = u.get_confirmation_token('foo@bar.com', force=True)
-        assert_equal(token, '54321')
-
-    # Some old users will not have an 'expired' key in their email_verifications.
-    # Assume the token in expired
-    def test_get_confirmation_token_if_email_verification_doesnt_have_expiration(self):
-        u = UserFactory()
-
-        email = fake.email()
-        u.add_unconfirmed_email(email)
-        # manually remove 'expiration' key
-        token = u.get_confirmation_token(email)
-        del u.email_verifications[token]['expiration']
-        u.save()
-
-        with assert_raises(ExpiredTokenError):
-            u.get_confirmation_token(email)
-
-    @mock.patch('website.security.random_string')
-    def test_get_confirmation_url(self, random_string):
-        random_string.return_value = 'abcde'
-        u = UserFactory()
-        u.add_unconfirmed_email('foo@bar.com')
-        assert_equal(u.get_confirmation_url('foo@bar.com'),
-                '{0}confirm/{1}/{2}/'.format(settings.DOMAIN, u._primary_key, 'abcde'))
-
-    @mock.patch('website.security.random_string')
-    def test_get_confirmation_url_for_external_service(self, random_string):
-        random_string.return_value = 'abcde'
-        u = UnconfirmedUserFactory()
-        assert_equal(u.get_confirmation_url(u.username, external_id_provider='service', destination='dashboard'),
-                '{0}confirm/external/{1}/{2}/?destination={3}'.format(settings.DOMAIN, u._id, 'abcde', 'dashboard'))
-
-    def test_get_confirmation_url_when_token_is_expired_raises_error(self):
-        u = UserFactory()
-        # Make sure token is already expired
-        expiration = timezone.now() - datetime.timedelta(seconds=1)
-        u.add_unconfirmed_email('foo@bar.com', expiration=expiration)
-
-        with assert_raises(ExpiredTokenError):
-            u.get_confirmation_url('foo@bar.com')
-
-    @mock.patch('website.security.random_string')
-    def test_get_confirmation_url_when_token_is_expired_force(self, random_string):
-        random_string.return_value = '12345'
-        u = UserFactory()
-        # Make sure token is already expired
-        expiration = timezone.now() - datetime.timedelta(seconds=1)
-        u.add_unconfirmed_email('foo@bar.com', expiration=expiration)
-
-        # sanity check
-        with assert_raises(ExpiredTokenError):
-            u.get_confirmation_token('foo@bar.com')
-
-        random_string.return_value = '54321'
-
-        url = u.get_confirmation_url('foo@bar.com', force=True)
-        expected = '{0}confirm/{1}/{2}/'.format(settings.DOMAIN, u._primary_key, '54321')
-        assert_equal(url, expected)
-
-    def test_confirm_primary_email(self):
-        u = UnconfirmedUserFactory()
-        token = u.get_confirmation_token(u.username)
-        confirmed = u.confirm_email(token)
-        u.save()
-        assert_true(confirmed)
-        assert_equal(len(u.email_verifications.keys()), 0)
-        assert_in(u.username, u.emails)
-        assert_true(u.is_registered)
-        assert_true(u.is_claimed)
-
-    def test_verify_confirmation_token(self):
-        u = UserFactory.build()
-        u.add_unconfirmed_email('foo@bar.com')
-        u.save()
-
-        with assert_raises(auth_exc.InvalidTokenError):
-            u.get_unconfirmed_email_for_token('badtoken')
-
-        valid_token = u.get_confirmation_token('foo@bar.com')
-        assert_true(u.get_unconfirmed_email_for_token(valid_token))
-        manual_expiration = timezone.now() - datetime.timedelta(0, 10)
-        u.email_verifications[valid_token]['expiration'] = manual_expiration
-
-        with assert_raises(auth_exc.ExpiredTokenError):
-            u.get_unconfirmed_email_for_token(valid_token)
-
-    def test_verify_confirmation_token_when_token_has_no_expiration(self):
-        # A user verification token may not have an expiration
-        email = fake.email()
-        u = UserFactory.build()
-        u.add_unconfirmed_email(email)
-        token = u.get_confirmation_token(email)
-        # manually remove expiration to simulate legacy user
-        del u.email_verifications[token]['expiration']
-        u.save()
-
-        assert_true(u.get_unconfirmed_email_for_token(token))
-
-    def test_factory(self):
-        # Clear users
-        Node.remove()
-        User.remove()
-        user = UserFactory()
-        assert_equal(User.find().count(), 1)
-        assert_true(user.username)
-        another_user = UserFactory(username='joe@mail.com')
-        assert_equal(another_user.username, 'joe@mail.com')
-        assert_equal(User.find().count(), 2)
-        assert_true(user.date_registered)
-
-    def test_format_surname(self):
-        user = UserFactory(fullname='Duane Johnson')
-        summary = user.get_summary(formatter='surname')
-        assert_equal(
-            summary['user_display_name'],
-            'Johnson'
-        )
-
-    def test_format_surname_one_name(self):
-        user = UserFactory(fullname='Rock')
-        summary = user.get_summary(formatter='surname')
-        assert_equal(
-            summary['user_display_name'],
-            'Rock'
-        )
-
-    def test_is_watching(self):
-        # User watches a node
-        watched_node = NodeFactory()
-        unwatched_node = NodeFactory()
-        config = WatchConfigFactory(node=watched_node)
-        self.user.watched.append(config)
-        self.user.save()
-        assert_true(self.user.is_watching(watched_node))
-        assert_false(self.user.is_watching(unwatched_node))
-
-    def test_serialize(self):
-        d = self.user.serialize()
-        assert_equal(d['id'], str(self.user._primary_key))
-        assert_equal(d['fullname'], self.user.fullname)
-        assert_equal(d['registered'], self.user.is_registered)
-        assert_equal(d['url'], self.user.url)
-
-    def test_set_password(self):
-        user = User(username=fake.email(), fullname='Nick Cage')
-        user.set_password('ghostrider')
-        user.save()
-        assert_true(check_password_hash(user.password, 'ghostrider'))
-
-    def test_check_password(self):
-        user = User(username=fake.email(), fullname='Nick Cage')
-        user.set_password('ghostrider')
-        user.save()
-        assert_true(user.check_password('ghostrider'))
-        assert_false(user.check_password('ghostride'))
-
-    def test_change_password(self):
-        old_password = 'password'
-        new_password = 'new password'
-        confirm_password = new_password
-        self.user.set_password(old_password)
-        self.user.save()
-        self.user.change_password(old_password, new_password, confirm_password)
-        assert_true(self.user.check_password(new_password))
-
-    def test_change_password_invalid(self, old_password=None, new_password=None, confirm_password=None,
-                                     error_message='Old password is invalid'):
-        self.user.set_password('password')
-        self.user.save()
-        with assert_raises(ChangePasswordError) as error:
-            self.user.change_password(old_password, new_password, confirm_password)
-            self.user.save()
-        assert_in(error_message, error.exception.message)
-        assert_false(self.user.check_password(new_password))
-
-    def test_change_password_invalid_old_password(self):
-        self.test_change_password_invalid(
-            'invalid old password',
-            'new password',
-            'new password',
-            'Old password is invalid',
-        )
-
-    def test_change_password_invalid_new_password_length(self):
-        self.test_change_password_invalid(
-            'password',
-            '12345',
-            '12345',
-            'Password should be at least eight characters',
-        )
-
-    def test_change_password_invalid_confirm_password(self):
-        self.test_change_password_invalid(
-            'password',
-            'new password',
-            'invalid confirm password',
-            'Password does not match the confirmation',
-        )
-
-    def test_change_password_invalid_blank_password(self, old_password='', new_password='', confirm_password=''):
-        self.test_change_password_invalid(
-            old_password,
-            new_password,
-            confirm_password,
-            'Passwords cannot be blank',
-        )
-
-    def test_change_password_invalid_blank_new_password(self):
-        for password in (None, '', '      '):
-            self.test_change_password_invalid_blank_password('password', password, 'new password')
-
-    def test_change_password_invalid_blank_confirm_password(self):
-        for password in (None, '', '      '):
-            self.test_change_password_invalid_blank_password('password', 'new password', password)
-
-    def test_url(self):
-        assert_equal(
-            self.user.url,
-            '/{0}/'.format(self.user._primary_key)
-        )
-
-    def test_absolute_url(self):
-        assert_equal(
-            self.user.absolute_url,
-            urlparse.urljoin(settings.DOMAIN, '/{0}/'.format(self.user._primary_key))
-        )
-
-    def test_profile_image_url(self):
-        expected = filters.gravatar(
-            self.user,
-            use_ssl=True,
-            size=settings.PROFILE_IMAGE_MEDIUM
-        )
-        assert_equal(self.user.profile_image_url(settings.PROFILE_IMAGE_MEDIUM), expected)
-
-    def test_profile_image_url_has_no_default_size(self):
-        expected = filters.gravatar(
-            self.user,
-            use_ssl=True,
-        )
-        assert_equal(self.user.profile_image_url(), expected)
-        size = urlparse.parse_qs(urlparse.urlparse(self.user.profile_image_url()).query).get('size')
-        assert_equal(size, None)
-
-    def test_activity_points(self):
-        assert_equal(self.user.get_activity_points(db=self.db),
-                    get_total_activity_count(self.user._primary_key))
-
-    def test_recently_added(self):
-        # Project created
-        project = ProjectFactory(creator=self.user)
-
-        assert_true(hasattr(self.user, 'recently_added'))
-
-        # Two users added as contributors
-        user2 = UserFactory()
-        user3 = UserFactory()
-        project.add_contributor(contributor=user2, auth=self.auth)
-        project.add_contributor(contributor=user3, auth=self.auth)
-        assert_equal(user3, self.user.recently_added[0])
-        assert_equal(user2, self.user.recently_added[1])
-        assert_equal(len(self.user.recently_added), 2)
-
-    def test_recently_added_multi_project(self):
-        # Three users are created
-        user2 = UserFactory()
-        user3 = UserFactory()
-        user4 = UserFactory()
-
-        # 2 projects created
-        project = ProjectFactory(creator=self.user)
-        project2 = ProjectFactory(creator=self.user)
-
-        # Users 2 and 3 are added to original project
-        project.add_contributor(contributor=user2, auth=self.auth)
-        project.add_contributor(contributor=user3, auth=self.auth)
-
-        # Users 2 and 3 are added to original project
-        project2.add_contributor(contributor=user2, auth=self.auth)
-        project2.add_contributor(contributor=user4, auth=self.auth)
-
-        assert_equal(user4, self.user.recently_added[0])
-        assert_equal(user2, self.user.recently_added[1])
-        assert_equal(user3, self.user.recently_added[2])
-        assert_equal(len(self.user.recently_added), 3)
-
-    def test_recently_added_length(self):
-        # Project created
-        project = ProjectFactory(creator=self.user)
-
-        assert_equal(len(self.user.recently_added), 0)
-        # Add 17 users
-        for _ in range(17):
-            project.add_contributor(
-                contributor=UserFactory(),
-                auth=self.auth
-            )
-
-        assert_equal(len(self.user.recently_added), 15)
-
-    def test_display_full_name_registered(self):
-        u = UserFactory()
-        assert_equal(u.display_full_name(), u.fullname)
-
-    def test_display_full_name_unregistered(self):
-        name = fake.name()
-        u = UnregUserFactory()
-        project = ProjectFactory()
-        project.add_unregistered_contributor(fullname=name, email=u.username,
-            auth=Auth(project.creator))
-        project.save()
-        assert_equal(u.display_full_name(node=project), name)
-
-    def test_get_projects_in_common(self):
-        user2 = UserFactory()
-        project = ProjectFactory(creator=self.user)
-        project.add_contributor(contributor=user2, auth=self.auth)
-        project.save()
-
-        project_keys = set([node._id for node in self.user.contributed])
-        projects = set(self.user.contributed)
-        user2_project_keys = set([node._id for node in user2.contributed])
-
-        assert_equal(self.user.get_projects_in_common(user2, primary_keys=True),
-                     project_keys.intersection(user2_project_keys))
-        assert_equal(self.user.get_projects_in_common(user2, primary_keys=False),
-                     projects.intersection(user2.contributed))
-
-    def test_n_projects_in_common(self):
-        user2 = UserFactory()
-        user3 = UserFactory()
-        project = ProjectFactory(creator=self.user)
-
-        project.add_contributor(contributor=user2, auth=self.auth)
-        project.save()
-
-        assert_equal(self.user.n_projects_in_common(user2), 1)
-        assert_equal(self.user.n_projects_in_common(user3), 0)
-
-    def test_user_get_cookie(self):
-        user = UserFactory()
-        super_secret_key = 'children need maps'
-        signer = itsdangerous.Signer(super_secret_key)
-        session = Session(data={
-            'auth_user_id': user._id,
-            'auth_user_username': user.username,
-            'auth_user_fullname': user.fullname,
-        })
-        session.save()
-
-        assert_equal(signer.unsign(user.get_or_create_cookie(super_secret_key)), session._id)
-
-    def test_user_get_cookie_no_session(self):
-        user = UserFactory()
-        super_secret_key = 'children need maps'
-        signer = itsdangerous.Signer(super_secret_key)
-        assert_equal(
-            0,
-            Session.find(Q('data.auth_user_id', 'eq', user._id)).count()
-        )
-
-        cookie = user.get_or_create_cookie(super_secret_key)
-
-        session = Session.find(Q('data.auth_user_id', 'eq', user._id))[0]
-
-        assert_equal(session._id, signer.unsign(cookie))
-        assert_equal(session.data['auth_user_id'], user._id)
-        assert_equal(session.data['auth_user_username'], user.username)
-        assert_equal(session.data['auth_user_fullname'], user.fullname)
-
-    def test_get_user_by_cookie(self):
-        user = UserFactory()
-        cookie = user.get_or_create_cookie()
-        assert_equal(user, User.from_cookie(cookie))
-
-    def test_get_user_by_cookie_returns_none(self):
-        assert_equal(None, User.from_cookie(''))
-
-    def test_get_user_by_cookie_bad_cookie(self):
-        assert_equal(None, User.from_cookie('foobar'))
-
-    def test_get_user_by_cookie_no_user_id(self):
-        user = UserFactory()
-        cookie = user.get_or_create_cookie()
-        session = Session.find_one(Q('data.auth_user_id', 'eq', user._id))
-        del session.data['auth_user_id']
-        assert_in('data', session.save())
-
-        assert_equal(None, User.from_cookie(cookie))
-
-    def test_get_user_by_cookie_no_session(self):
-        user = UserFactory()
-        cookie = user.get_or_create_cookie()
-        Session.remove()
-        assert_equal(
-            0,
-            Session.find(Q('data.auth_user_id', 'eq', user._id)).count()
-        )
-        assert_equal(None, User.from_cookie(cookie))
-
-
-class TestUserParse(unittest.TestCase):
-
-    def test_parse_first_last(self):
-        parsed = impute_names_model('John Darnielle')
-        assert_equal(parsed['given_name'], 'John')
-        assert_equal(parsed['family_name'], 'Darnielle')
-
-    def test_parse_first_last_particles(self):
-        parsed = impute_names_model('John van der Slice')
-        assert_equal(parsed['given_name'], 'John')
-        assert_equal(parsed['family_name'], 'van der Slice')
-
-
-class TestDisablingUsers(OsfTestCase):
-    def setUp(self):
-        super(TestDisablingUsers, self).setUp()
-        self.user = UserFactory()
-
-    def test_user_enabled_by_default(self):
-        assert_false(self.user.is_disabled)
-
-    def test_disabled_user(self):
-        """Ensure disabling a user sets date_disabled"""
-        self.user.is_disabled = True
-        self.user.save()
-
-        assert_true(isinstance(self.user.date_disabled, datetime.datetime))
-        assert_true(self.user.is_disabled)
-        assert_false(self.user.is_active)
-
-    def test_reenabled_user(self):
-        """Ensure restoring a disabled user unsets date_disabled"""
-        self.user.is_disabled = True
-        self.user.save()
-
-        self.user.is_disabled = False
-        self.user.save()
-
-        assert_is_none(self.user.date_disabled)
-        assert_false(self.user.is_disabled)
-        assert_true(self.user.is_active)
-
-    def test_is_disabled_idempotency(self):
-        self.user.is_disabled = True
-        self.user.save()
-
-        old_date_disabled = self.user.date_disabled
-
-        self.user.is_disabled = True
-        self.user.save()
-
-        new_date_disabled = self.user.date_disabled
-
-        assert_equal(new_date_disabled, old_date_disabled)
-
-    @mock.patch('website.mailchimp_utils.get_mailchimp_api')
-    def test_disable_account_and_remove_sessions(self, mock_mail):
-        session1 = SessionFactory(user=self.user, date_created=(timezone.now() - datetime.timedelta(seconds=settings.OSF_SESSION_TIMEOUT)))
-        session2 = SessionFactory(user=self.user, date_created=(timezone.now() - datetime.timedelta(seconds=settings.OSF_SESSION_TIMEOUT)))
-
-        self.user.mailchimp_mailing_lists[settings.MAILCHIMP_GENERAL_LIST] = True
-        self.user.save()
-        self.user.disable_account()
-
-        assert_true(self.user.is_disabled)
-        assert_true(isinstance(self.user.date_disabled, datetime.datetime))
-        assert_false(self.user.mailchimp_mailing_lists[settings.MAILCHIMP_GENERAL_LIST])
-
-        assert_false(Session.load(session1._id))
-        assert_false(Session.load(session2._id))
-
-    def test_disable_account_api(self):
-        settings.ENABLE_EMAIL_SUBSCRIPTIONS = True
-        with assert_raises(mailchimp_utils.mailchimp.InvalidApiKeyError):
-            self.user.disable_account()
-
-
-class TestMergingUsers(OsfTestCase):
-
-    def setUp(self):
-        super(TestMergingUsers, self).setUp()
-        with self.context:
-            handlers.celery_before_request()
-
-        self.master = UserFactory(
-            fullname='Joe Shmo',
-            is_registered=True,
-            emails=['joe@mail.com'],
-        )
-        self.dupe = UserFactory(
-            fullname='Joseph Shmo',
-            emails=['joseph123@hotmail.com']
-        )
-
-    def _merge_dupe(self):
-        '''Do the actual merge.'''
-        self.master.merge_user(self.dupe)
-        self.master.save()
-
-    def test_bookmark_collection_nodes_arent_merged(self):
-        dashnode = find_bookmark_collection(self.dupe)
-
-        self._merge_dupe()
-
-        assert_not_in(dashnode, self.master. contributed)
-
-    def test_dupe_is_merged(self):
-        self._merge_dupe()
-        assert_true(self.dupe.is_merged)
-        assert_equal(self.dupe.merged_by, self.master)
-
-    def test_dupe_email_is_appended(self):
-        self._merge_dupe()
-        assert_in('joseph123@hotmail.com', self.master.emails)
-
-    def test_send_user_merged_signal(self):
-        self.dupe.mailchimp_mailing_lists['foo'] = True
-        self.dupe.save()
-
-        with capture_signals() as mock_signals:
-            self._merge_dupe()
-            assert_equal(mock_signals.signals_sent(), set([user_merged]))
-
-    @mock.patch('website.mailchimp_utils.get_mailchimp_api')
-    def test_merged_user_unsubscribed_from_mailing_lists(self, mock_get_mailchimp_api):
-        list_name = 'foo'
-        username = self.dupe.username
-        self.dupe.mailchimp_mailing_lists[list_name] = True
-        self.dupe.save()
-        mock_client = mock.MagicMock()
-        mock_get_mailchimp_api.return_value = mock_client
-        mock_client.lists.list.return_value = {'data': [{'id': 2, 'list_name': list_name}]}
-        list_id = mailchimp_utils.get_list_id_from_name(list_name)
-        self._merge_dupe()
-        handlers.celery_teardown_request()
-        mock_client.lists.unsubscribe.assert_called_with(id=list_id, email={'email': username}, send_goodbye=False)
-        assert_false(self.dupe.mailchimp_mailing_lists[list_name])
-
-    def test_inherits_projects_contributed_by_dupe(self):
-        project = ProjectFactory()
-        project.add_contributor(self.dupe)
-        project.save()
-        self._merge_dupe()
-        project.reload()
-        assert_true(project.is_contributor(self.master))
-        assert_false(project.is_contributor(self.dupe))
-
-    def test_inherits_projects_created_by_dupe(self):
-        project = ProjectFactory(creator=self.dupe)
-        self._merge_dupe()
-        project.reload()
-        assert_equal(project.creator, self.master)
-
-    def test_adding_merged_user_as_contributor_adds_master(self):
-        project = ProjectFactory(creator=UserFactory())
-        self._merge_dupe()
-        project.add_contributor(contributor=self.dupe)
-        assert_true(project.is_contributor(self.master))
-        assert_false(project.is_contributor(self.dupe))
-
-    def test_merging_dupe_who_is_contributor_on_same_projects(self):
-        # Both master and dupe are contributors on the same project
-        project = ProjectFactory()
-        project.add_contributor(contributor=self.master)
-        project.add_contributor(contributor=self.dupe)
-        project.save()
-        self._merge_dupe()  # perform the merge
-        project.reload()
-        assert_true(project.is_contributor(self.master))
-        assert_false(project.is_contributor(self.dupe))
-        assert_equal(len(project.contributors), 2) # creator and master
-                                                   # are the only contribs
 
 
 class TestGUID(OsfTestCase):
@@ -4341,7 +3346,6 @@ class TestNodeLog(OsfTestCase):
 
         created_log = project.logs[0]
         assert_false(created_log.can_view(unrelated, Auth(user=project.creator)))
-
 
     def test_original_node_and_current_node_for_registration_logs(self):
         user = UserFactory()
